requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.34.34
botocore==1.34.162
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.109.2
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
openai==1.97.1
passlib==1.7.4
propcache==0.3.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
requests==2.32.4
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.36.3
tqdm==4.67.1
twilio==8.12.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.27.1
yarl==1.20.1
opensearch-py==3.0.0